% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_enrichment.R
\name{flx.plot_labeling_enrichment}
\alias{flx.plot_labeling_enrichment}
\title{Plot labeling enrichment}
\usage{
flx.plot_labeling_enrichment(
  listed.data = NULL,
  compound = NULL,
  enrichment_lower_bound = 0,
  sample_group = NULL,
  facet.nrow = 1,
  orderedSample = NULL
)
}
\arguments{
\item{listed.data}{A list of datasets. It is the output from the function of "flx.correct_natural_abundance". The two subsets "Corrected" and "Normalized" datasets are in particular used for plotting under the hood.}

\item{compound}{A character vector of length one to specify the compound name.}

\item{enrichment_lower_bound}{A single number to specify the lower boundary of the y-axis in the bar plot. By default, an enrichment range of 0 - 1 is plotted on the y-axis in the bar plot.}

\item{sample_group}{A dataset of sample information, optional argument. If specified, the plot is faceted into panels based on sample groups. The input dataset contains two essential columns: the specified column "sample" contains all sample names as contained in the "listed.data" argument input, and the column of "group" denotes the category by which the plot is faceted into group-panel. Extra columns can be stored in this dataset for other analysis, but will not be utilized for this function.}
}
\value{
A ggplot.
}
\description{
This function takes as input the output list from the function of "flx.correct_natural_abundance", and plot the labeling enrichment for any specified compound. When the "sample_group" argument is specified, the plot is faceted into panels based on the specified group category.
}
