% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_background.R
\name{flx.subtract_background}
\alias{flx.subtract_background}
\title{Background subtraction.}
\usage{
flx.subtract_background(dataset, blankSamples = NA)
}
\arguments{
\item{dataset}{Dataframe. The first three columns are reserved to "isotopeLbel", "compound", and "formula", respectively. The following columns contain the sample signal intensities as numeric values; non-numeric values, such as characters or symbols, must be avoided.}

\item{blankSamples}{A vector of blank samples' names.It indicates the blank samples whose average signal should be subtracted from real sample signal.}
}
\value{
A data frame, with background subtracted. The blank samples (columns) are removed from the output dataset.
}
\description{
Background level of each isotopomer of each metabolite is averaged from designated blank sample(s), which are then subtracted from the associated signal intensity in biological samples.
}
